backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.4
python-dotenv==1.0.1
pinecone==6.0.2
google-genai
